reorgnize repo and finetune milestone done
Enter your instruction: how many times does r appear in strawberry
Generating response...

Response:
The first occurrence of R appears at index 0, which is the second character in the string "strawberry". The second occurrence of R appears at index 2, which is the second character in the string "r". The third occurrence of R appears at index 4, which is the second character in the string "r". Therefore, the total number of occurrences of R in the string is 3.